requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
pandas>=2.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
setuptools>=68.0.0
wheel>=0.41.0
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0
seaborn>=0.12.0
gradio>=3.39.0
plotly>=5.17.0
py-cpuinfo>=9.0.0
psutil>=5.9.0